requirements.txt
db-dtypes==1.5.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
openai==1.109.0
pandas==2.3.3
pyarrow==21.0.0
requests==2.32.5
tqdm==4.67.3
python-dateutil==2.9.0.post0
